Moved Create/Send Invitation from Event component to Event Details Component
