+++ source/ClientApp/src/app/events/events.component.html
@@ -92,10 +92,6 @@
           <mat-icon>delete_forever</mat-icon>
       </button>
 
-      <button mat-button (click)="loadInvite(event)" matTooltip="Email Guest List">
-          <mat-icon>email</mat-icon>
-    </button>
-
     </div>
   </div>
 </div>
